refactor: combine OG generators into single shared module
Merge default, hub, and workflow OG generators into one file with
shared base layout (OGBase), Header, Footer components, and common
constants. Adds gradient overlay to default OG image.
