+++ app/api/og/default/route.tsx
@@ -1,5 +1,5 @@
 // start custom keeperhub code //
-import { generateDefaultOGImage } from "@/keeperhub/api/og/default/generate-og-default";
+import { generateDefaultOGImage } from "@/keeperhub/api/og/generate-og";
 
 export function GET(): Response {
   return generateDefaultOGImage();
